feat: add SettingsProvider and useSettings hook for managing application settings
